+++ tests/python/openassetio/hostAPI/test_manager.py
@@ -25,16 +25,15 @@ from unittest import mock
 
 import pytest
 
-from openassetio import Context, TraitsData
+from openassetio import Context, TraitsData, managerAPI
 from openassetio.hostAPI import Manager
-from openassetio.managerAPI import HostSession
 
 
 ## @todo Remove comments regarding Entity methods when splitting them from core API
 
 @pytest.fixture
 def host_session():
-    return mock.create_autospec(HostSession)
+    return mock.create_autospec(managerAPI.HostSession)
 
 
 @pytest.fixture
@@ -479,14 +478,14 @@ class Test_Manager_createContext:
 
     def test_context_is_created_with_expected_properties(self, manager, mock_manager_interface):
 
-        state_a = "state-a"
+        state_a = managerAPI.ManagerStateBase()
         mock_manager_interface.mock.createState.return_value = state_a
 
         context_a = manager.createContext()
 
-        assert context_a.access == Context.kRead
+        assert context_a.access == Context.kUnknown
         assert context_a.retention == Context.kTransient
-        assert context_a.managerInterfaceState is state_a
+        assert context_a.managerState is state_a
         assert context_a.locale is None
         mock_manager_interface.mock.createState.assert_called_once()
 
@@ -494,7 +493,7 @@ class Test_Manager_createContext:
             self, manager, mock_manager_interface):
         # pylint: disable=protected-access
 
-        state_a = "state-a"
+        state_a = managerAPI.ManagerStateBase()
         mock_manager_interface.mock.createState.return_value = state_a
         context_a = manager.createContext()
         context_a.access = Context.kWrite
@@ -502,14 +501,14 @@ class Test_Manager_createContext:
         context_a.locale = TraitsData()
         mock_manager_interface.mock.reset_mock()
 
-        state_b = "state-b"
+        state_b = managerAPI.ManagerStateBase()
         mock_manager_interface.mock.createState.return_value = state_b
         a_host_session = manager._Manager__hostSession
 
         context_b = manager.createContext(parent=context_a)
 
         assert context_b is not context_a
-        assert context_b.managerInterfaceState is state_b
+        assert context_b.managerState is state_b
         assert context_b.access == context_a.access
         assert context_b.retention == context_a.retention
         assert context_b.locale == context_b.locale
@@ -525,9 +524,9 @@ class Test_Manager_freezeContext:
         expected_token = "a_frozen_token"
         mock_manager_interface.mock.freezeState.return_value = expected_token
 
-        initial_state = "initial_state"
+        initial_state = managerAPI.ManagerStateBase()
         a_context = Context()
-        a_context.managerInterfaceState = initial_state
+        a_context.managerState = initial_state
 
         actual_token = manager.freezeContext(a_context)
 
@@ -544,13 +543,13 @@ class Test_Manager_thawContext:
     def test_when_called_then_the_managers_thawed_state_is_set_in_the_context(
              self, manager, mock_manager_interface):
 
-        expected_state = "thawed_state"
+        expected_state = managerAPI.ManagerStateBase()
         mock_manager_interface.mock.thawState.return_value = expected_state
 
         a_token = "frozen_token"
         a_context = manager.thawContext(a_token)
 
-        assert a_context.managerInterfaceState is expected_state
+        assert a_context.managerState is expected_state
 
         # pylint: disable=protected-access
         a_host_session = manager._Manager__hostSession